Refactor filetree
