+++ src/dict/main/mod.rs
@@ -1,6 +1,5 @@
 //! Main dictionary.
 
-pub mod html;
 mod ir;
 mod locale;
 mod yomitan;
