+++ src/dict/writer/mdict_text/renderer.rs
@@ -1,0 +1,6 @@
+use crate::dict::writer::renderer::Renderer;
+
+// Use the default. Goldendict supports all the html features.
+pub struct MdictTextRenderer;
+
+impl Renderer for MdictTextRenderer {}
