+++ src/dict/writer/mod.rs
@@ -1,26 +1,25 @@
-use std::{
-    collections::HashSet,
-    fmt,
-    fs::{self, File},
-    io::{BufWriter, Write},
-    path::Path,
-};
+use std::fmt;
 
 use anyhow::Result;
 use clap::ValueEnum;
 
 use crate::{
     cli::{LangSpecs, Options},
-    dict::{
-        Dictionary, Intermediate,
-        main::html::{BaseRenderer, Renderer, StardictRenderer},
-    },
-    models::yomitan::{DetailedDefinition, YomitanDict, YomitanEntry},
+    dict::{Dictionary, Intermediate},
     path::PathManager,
 };
 
-// mod html;
+mod debug_forms;
+mod html;
+mod mdict_text;
+mod renderer;
+mod stardict;
 mod yomitan;
+
+use debug_forms::write_debug_forms;
+use html::write_html;
+use mdict_text::write_mdict_text;
+use stardict::write_stardict;
 use yomitan::{write_yomitan, write_yomitan_simple};
 
 #[derive(ValueEnum, Debug, Default, Clone, Copy)]
@@ -94,323 +93,4 @@ impl WriterFormat {
             Self::Skip => Ok(()),
         }
     }
-}
-
-fn write_debug_forms(_: &Options, pm: &PathManager, ydict: YomitanDict) -> Result<()> {
-    let dname = pm.dict_name_expanded();
-    let filepath = format!("html/test-{}.txt", dname);
-    let filename = Path::new(&filepath);
-    if let Some(parent) = filename.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    let file = File::create(filename)?;
-    let mut writer = BufWriter::new(file);
-
-    // NOTE: we are undoing work and going back to irs... but we can't replace
-    // the yomitan dict with irs, since then we should assume that the irs can be
-    // of ANY dictionary (and we only care about the main one)
-    // ... and in theory, there is no guarantee that the irs format of the main
-    // dictionary won't change, while this logic remains the same.
-    let mut grouped_by: HashMap<String, Vec<String>> = HashMap::new();
-
-    for entry in ydict.term_info_form {
-        let term = entry.term.to_string();
-        for def in entry.definitions {
-            let DetailedDefinition::Inflection((from, _tags)) = def else {
-                panic!("forms must be made from inflections");
-            };
-            let tos = grouped_by.entry(from).or_default();
-            if !tos.contains(&term) {
-                tos.push(term.clone());
-            }
-        }
-    }
-
-    for (from, tos) in &grouped_by {
-        writer.write_all(from.as_bytes())?;
-        writer.write_all(b"\n")?;
-        for to in tos {
-            writer.write_all(to.as_bytes())?;
-            writer.write_all(b"\n")?;
-        }
-        writer.write_all(b"\n")?;
-    }
-
-    crate::utils::pretty_println_at_path("Wrote debug file", filename);
-
-    Ok(())
-}
-
-// TODO: move this
-fn write_html(opts: &Options, pm: &PathManager, ydict: YomitanDict) -> Result<()> {
-    let dname = pm.dict_name_expanded();
-    let filepath = format!("html/test-{}.html", dname);
-    let filename = Path::new(&filepath);
-    if let Some(parent) = filename.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    let file = File::create(filename)?;
-    let mut writer = BufWriter::new(file);
-
-    writer.write_all(
-        br#"<!DOCTYPE html>
-            <html>
-            <head>
-            <meta charset="utf-8">
-            <link rel="stylesheet" href="styles.css">
-            </head>
-            <body>
-            "#,
-    )?;
-
-    for entry in ydict.into_iter_flat() {
-        let html = BaseRenderer::render_entry(&entry).into_string();
-        if opts.pretty {
-            let pretty = prettify_html(&html);
-            writer.write_all(pretty.as_bytes())?;
-        } else {
-            writer.write_all(html.as_bytes())?;
-        }
-    }
-    writer.write_all(br#"</body></html>"#)?;
-    crate::utils::pretty_println_at_path("Wrote file", filename);
-
-    Ok(())
-}
-
-// TODO: move this
-fn write_mdict_text(opts: &Options, pm: &PathManager, ydict: YomitanDict) -> Result<()> {
-    let dname = pm.dict_name_expanded();
-    let filepath = format!("html/test-{}.txt", dname);
-    let filename = Path::new(&filepath);
-    if let Some(parent) = filename.parent() {
-        std::fs::create_dir_all(parent)?;
-    }
-    let file = File::create(filename)?;
-    let mut writer = BufWriter::new(file);
-
-    for entry in ydict.into_iter_flat() {
-        writer.write_all(entry.term().as_bytes())?;
-        writer.write_all(b"\n")?;
-
-        let html = BaseRenderer::render_entry(&entry).into_string();
-        // Requires a css include in each entry!
-        writer.write_all(b"<link rel='stylesheet' href='styles.css' type='text/css'>")?;
-        if opts.pretty {
-            let pretty = prettify_html(&html);
-            writer.write_all(pretty.as_bytes())?;
-        } else {
-            writer.write_all(html.as_bytes())?;
-        }
-        writer.write_all(b"\n")?;
-        writer.write_all(b"</>")?;
-        writer.write_all(b"\n")?;
-    }
-    writer.write_all(br#"</body></html>"#)?;
-    crate::utils::pretty_println_at_path("Wrote html", filename);
-
-    Ok(())
-}
-
-// This is scuffed and introduces whitespace where it shouldn't, which then
-// makes the css white-space: pre-line do some insane mangling.
-// Stick with this for tests, without that css line, but remember to add it
-// back in the final version.
-// TODO: move this
-fn prettify_html(html: &str) -> String {
-    let mut result = String::new();
-    let mut indent: usize = 0;
-    let mut in_tag = false;
-
-    for ch in html.chars() {
-        match ch {
-            '<' => {
-                if !in_tag {
-                    result.push('\n');
-                    result.push_str(&"  ".repeat(indent)); // main offender
-                }
-                in_tag = true;
-                result.push(ch);
-            }
-            '>' => {
-                result.push(ch);
-                in_tag = false;
-
-                // Adjust indent based on tag type
-                let last_tag = result.split('<').last().unwrap_or("");
-                if last_tag.starts_with('/') {
-                    indent = indent.saturating_sub(1);
-                } else if !last_tag.ends_with('/') && !last_tag.starts_with('!') {
-                    indent += 1;
-                }
-            }
-            _ => result.push(ch),
-        }
-    }
-
-    result
-}
-
-use std::collections::HashMap;
-
-type Definitions = Vec<(String, String)>;
-type Synonyms = Vec<(String, Vec<String>)>;
-
-fn write_stardict(_: &Options, _: &PathManager, ydict: YomitanDict) -> Result<()> {
-    let term_info_erased: Vec<_> = ydict
-        .term_info
-        .into_iter()
-        .map(YomitanEntry::TermInfo)
-        .collect();
-    let term_meta_erased: Vec<_> = ydict
-        .term_meta
-        .into_iter()
-        .map(YomitanEntry::TermMeta)
-        .collect();
-
-    let mut lemmas = term_info_erased;
-    lemmas.extend(term_meta_erased);
-    let forms = ydict.term_info_form;
-
-    let opath = Path::new("wty-stardict");
-    let _ = fs::create_dir(&opath);
-
-    let definitions = lemmas
-        .into_iter()
-        // the into below is for PreEscapedString > String
-        .map(|entry| {
-            (
-                entry.term().to_string(),
-                StardictRenderer::render_entry(&entry).into(),
-            )
-        })
-        .collect();
-    let synonyms = forms
-        .into_iter()
-        .map(|entry| {
-            let redirects_to = entry.term;
-            let redirects_from = {
-                entry
-                    .definitions
-                    .into_iter()
-                    .map(|def| {
-                        let DetailedDefinition::Inflection((from, _tags)) = def else {
-                            panic!("forms must be made from inflections");
-                        };
-                        from
-                    })
-                    .collect::<Vec<_>>()
-            };
-            (redirects_to, redirects_from)
-        })
-        .collect();
-    write_to_stardict(opath, definitions, synonyms)?;
-    Ok(())
-}
-
-// In my machine (ubuntu, flatpak), dict should be @
-// ~/.var/app/rocks.koreader.KOReader/config/koreader/data/dict
-fn write_to_stardict(
-    opath: &Path,
-    mut definitions: Definitions,
-    mut synonyms: Synonyms,
-) -> Result<()> {
-    let output_base = opath.join("output");
-
-    let dict_path = output_base.with_extension("dict");
-    let idx_path = output_base.with_extension("idx");
-    let syn_path = output_base.with_extension("syn");
-    let ifo_path = output_base.with_extension("ifo");
-
-    definitions.sort_by(|a, b| a.0.to_lowercase().cmp(&b.0.to_lowercase()));
-    synonyms.sort_by(|a, b| a.0.to_lowercase().cmp(&b.0.to_lowercase()));
-
-    let word_to_index: HashMap<&str, u32> = definitions
-        .iter()
-        .enumerate()
-        .map(|(i, (word, _))| (word.as_str(), i as u32))
-        .collect();
-
-    // Write .dict and build index
-    let mut index_data: Vec<Vec<u8>> = Vec::new();
-    let mut offset: u32 = 0;
-
-    {
-        let dict_file = File::create(&dict_path)?;
-        let mut dict_writer = BufWriter::new(dict_file);
-
-        for (word, definition) in &definitions {
-            let encoded_def = definition.as_bytes();
-            dict_writer.write_all(encoded_def)?;
-            let length = encoded_def.len() as u32;
-
-            let mut entry = Vec::new();
-            entry.extend_from_slice(word.as_bytes());
-            entry.push(0u8); // null terminator
-            entry.extend_from_slice(&offset.to_be_bytes());
-            entry.extend_from_slice(&length.to_be_bytes());
-
-            index_data.push(entry);
-            offset += length;
-        }
-    }
-
-    // Write .idx
-    {
-        let idx_file = File::create(&idx_path)?;
-        let mut idx_writer = BufWriter::new(idx_file);
-        for item in &index_data {
-            idx_writer.write_all(item)?;
-        }
-    }
-
-    // Write .syn
-    let mut syn_count: u32 = 0;
-    {
-        let syn_file = File::create(&syn_path)?;
-        let mut syn_writer = BufWriter::new(syn_file);
-        // We don't want to add redirections twice
-        let mut seen = HashSet::new();
-        // alias redirects to target!
-        for (alias, targets) in &synonyms {
-            for target in targets {
-                let Some(&idx) = word_to_index.get(target.as_str()) else {
-                    continue;
-                };
-                if !seen.insert((alias, idx)) {
-                    continue;
-                }
-                // tracing::error!("Added {alias} | {target}");
-                syn_writer.write_all(alias.as_bytes())?;
-                syn_writer.write_all(&[0u8])?;
-                syn_writer.write_all(&idx.to_be_bytes())?;
-                syn_count += 1;
-            }
-        }
-    }
-
-    // Write .ifo
-    let wordcount = definitions.len();
-    let idxfilesize = std::fs::metadata(&idx_path)?.len();
-    let bookname = "wty";
-
-    let ifo_content = format!(
-        "StarDict's dict ifo file\nversion=3.0.0\nbookname={bookname}\nwordcount={wordcount}\nsynwordcount={syn_count}\nidxfilesize={idxfilesize}\nsametypesequence=h\n"
-    );
-
-    {
-        let mut ifo_file = File::create(&ifo_path)?;
-        ifo_file.write_all(ifo_content.as_bytes())?;
-    }
-
-    println!("Done!");
-    println!("Definitions: {wordcount}");
-    println!("Synonyms: {syn_count}");
-    println!("Files created:");
-    println!(" - {}", dict_path.display());
-    println!(" - {}", idx_path.display());
-    println!(" - {}", syn_path.display());
-    println!(" - {}", ifo_path.display());
-
-    Ok(())
 }
